Code re-arrangement
